requirements.txt
sqlglot>=24.0.0
duckdb>=0.10.0
anthropic>=0.34.0
openai>=1.40.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0

